remove base row builder
